Basic lwIP for STM32H7 BSP
This patch updates the BSP to have basic lwIP support.
This includes a new section in the linker base file in case
lwIP is used for the STM32H7 BSP. This is not an ideal solution
but I have not found a better one.

Other adaptiosn include a new option to load lwIP and disabling
some preprocessor guards if the options is set to true.

I also did some smaller tweaks like providing another default
implementation for HAL_GetTick() and a small include fix