RTEMS_USE_LWIP option to use lwip in RTEMS
This adds a configuration option if lwIP is used instead of libbsd.
It works by specifying RTEMS_USE_LWIP in the config.ini file.

The option is still STM32H7 specific can could in principle be
reused for other BSPs. For the STM32H7, it allows the execution
of some operations necessary for lwIP to work properly.